Revert "refactor(user): organize use cases by operation"
This reverts commit d60a99af84993b86317e057a88c604403d7c1891.
