+++ src/main/java/com/ironcore/application/user/usecase/ChangePasswordCommand.java
@@ -1,4 +1,4 @@
-package com.ironcore.application.user.usecase.changepassword;
+package com.ironcore.application.user.usecase;
 
 import com.ironcore.domain.user.valueobject.RawPassword;
 import com.ironcore.domain.user.valueobject.UserId;
