+++ src/test/java/com/ironcore/application/user/usecase/BootstrapSingleUserUseCaseTest.java
@@ -2,8 +2,6 @@ package com.ironcore.application.user.usecase;
 
 import com.ironcore.application.exception.OperationNotAllowedException;
 import com.ironcore.application.user.service.PasswordHashingService;
-import com.ironcore.application.user.usecase.bootstrap.BootstrapSingleUserCommand;
-import com.ironcore.application.user.usecase.bootstrap.BootstrapSingleUserUseCase;
 import com.ironcore.domain.user.enums.SexType;
 import com.ironcore.domain.user.model.User;
 import com.ironcore.domain.user.repository.UserRepository;
