+++ src/test/java/com/ironcore/interfaces/rest/user/UserControllerTest.java
@@ -3,12 +3,7 @@ package com.ironcore.interfaces.rest.user;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.ironcore.application.exception.InvalidCredentialsException;
 import com.ironcore.application.logging.error.port.ErrorLogPublisher;
-import com.ironcore.application.user.usecase.changepassword.ChangePasswordCommand;
-import com.ironcore.application.user.usecase.changepassword.ChangePasswordUseCase;
-import com.ironcore.application.user.usecase.getauthenticateduser.GetAuthenticatedUserUseCase;
-import com.ironcore.application.user.usecase.getauthenticateduser.UserProfileResult;
-import com.ironcore.application.user.usecase.initialchangepassword.InitialChangePasswordCommand;
-import com.ironcore.application.user.usecase.initialchangepassword.InitialChangePasswordUseCase;
+import com.ironcore.application.user.usecase.*;
 import com.ironcore.domain.user.repository.UserRepository;
 import com.ironcore.domain.user.valueobject.Email;
 import com.ironcore.domain.user.valueobject.RawPassword;
